+++ src/main/java/io/nats/client/impl/NatsDispatcher.java
@@ -19,28 +19,40 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import io.nats.client.Subscription;
 import io.nats.client.Dispatcher;
 import io.nats.client.MessageHandler;
 
 class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
 
     private MessageQueue incoming;
-    private MessageHandler handler;
+    private MessageHandler defaultHandler;
 
     private Future<Boolean> thread;
     private final AtomicBoolean running;
 
     private String id;
 
-    private Map<String, NatsSubscription> subscriptions;
+    // We will use the subject as the key for subscriptions that use the
+    // default handler.
+    private Map<String, NatsSubscription> subscriptionsUsingDefaultHandler;
+    // We will use the SID as the key. Sicne these subscriptions provide
+    // their own handlers, we allow duplicates. There is a subtle but very
+    // important difference here.
+    private Map<String, NatsSubscription> subscriptionsWithHandlers;
+    // We use the SID as the key here.
+    private Map<String, MessageHandler> subscriptionHandlers;
+
     private Duration waitForMessage;
 
 
     NatsDispatcher(NatsConnection conn, MessageHandler handler) {
         super(conn);
-        this.handler = handler;
+        this.defaultHandler = handler;
         this.incoming = new MessageQueue(true);
-        this.subscriptions = new ConcurrentHashMap<>();
+        this.subscriptionsUsingDefaultHandler = new ConcurrentHashMap<>();
+        this.subscriptionsWithHandlers = new ConcurrentHashMap<>();
+        this.subscriptionHandlers = new ConcurrentHashMap<>();
         this.running = new AtomicBoolean(false);
         this.waitForMessage = Duration.ofMinutes(5); // This can be long since we aren't doing anything
     }
@@ -58,7 +70,7 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
     public void run() {
         try {
             while (this.running.get()) {
-                
+
                 NatsMessage msg = this.incoming.pop(this.waitForMessage);
 
                 if (msg == null) {
@@ -76,8 +88,14 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
                     sub.incrementDeliveredCount();
                     this.incrementDeliveredCount();
 
+                    MessageHandler currentHandler = this.defaultHandler;
+                    MessageHandler customHandler = this.subscriptionHandlers.get(sub.getSID());
+                    if (customHandler != null) {
+                        currentHandler = customHandler;
+                    }
+
                     try {
-                        handler.onMessage(msg);
+                        currentHandler.onMessage(msg);
                     } catch (Exception exp) {
                         this.connection.processException(exp);
                     }
@@ -117,11 +135,16 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
         }
 
         if (unsubscribeAll) {
-            this.subscriptions.forEach((subj, sub) -> {
+            this.subscriptionsUsingDefaultHandler.forEach((subj, sub) -> {
+                this.connection.unsubscribe(sub, -1);
+            });
+            this.subscriptionsWithHandlers.forEach((sid, sub) -> {
                 this.connection.unsubscribe(sub, -1);
             });
         } else {
-            this.subscriptions.clear();
+            this.subscriptionsUsingDefaultHandler.clear();
+            this.subscriptionsWithHandlers.clear();
+            this.subscriptionHandlers.clear();
         }
     }
 
@@ -138,14 +161,29 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
     }
 
     void resendSubscriptions() {
-        this.subscriptions.forEach((id, sub)->{
+        this.subscriptionsUsingDefaultHandler.forEach((id, sub)->{
+            this.connection.sendSubscriptionMessage(sub.getSID(), sub.getSubject(), sub.getQueueName(), true);
+        });
+        this.subscriptionsWithHandlers.forEach((sid, sub)->{
             this.connection.sendSubscriptionMessage(sub.getSID(), sub.getSubject(), sub.getQueueName(), true);
         });
     }
 
-    // Called by the connection when the subscription is removed
+    // Called by the connection when a subscription is removed.
+    // We will first attempt to remove from subscriptionsWithHandlers
+    // using the sub's SID, and if we don't find it there, we'll check
+    // the subscriptionsUsingDefaultHandler Map and verify the SID
+    // matches before removing. By verifying the SID in all cases we can
+    // be certain we're removing the correct Subscription.
     void remove(NatsSubscription sub) {
-        subscriptions.remove(sub.getSubject());
+        if (this.subscriptionsWithHandlers.remove(sub.getSID()) != null) {
+            this.subscriptionHandlers.remove(sub.getSID());
+        } else {
+            NatsSubscription s = this.subscriptionsUsingDefaultHandler.get(sub.getSubject());
+            if (s.getSID() == sub.getSID()) {
+                this.subscriptionsUsingDefaultHandler.remove(sub.getSubject());
+            }
+        }
     }
 
     public Dispatcher subscribe(String subject) {
@@ -153,7 +191,19 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
             throw new IllegalArgumentException("Subject is required in subscribe");
         }
 
-        return this.subscribeImpl(subject, null);
+        this.subscribeImpl(subject, null, null);
+        return this;
+    }
+
+    public Subscription subscribe(String subject, MessageHandler handler) {
+        if (subject == null || subject.length() == 0) {
+            throw new IllegalArgumentException("Subject is required in subscribe");
+        }
+
+        if (handler == null) {
+            throw new IllegalArgumentException("MessageHandler is required in subscribe");
+        }
+        return this.subscribeImpl(subject, null, handler);
     }
 
     public Dispatcher subscribe(String subject, String queueName) {
@@ -164,40 +214,68 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
         if (queueName == null || queueName.length() == 0) {
             throw new IllegalArgumentException("QueueName is required in subscribe");
         }
-        return this.subscribeImpl(subject, queueName);
+        this.subscribeImpl(subject, queueName, null);
+        return this;
     }
 
+    public Subscription subscribe(String subject, String queueName,  MessageHandler handler) {
+        if (subject == null || subject.length() == 0) {
+            throw new IllegalArgumentException("Subject is required in subscribe");
+        }
+
+        if (queueName == null || queueName.length() == 0) {
+            throw new IllegalArgumentException("QueueName is required in subscribe");
+        }
+
+        if (handler == null) {
+            throw new IllegalArgumentException("MessageHandler is required in subscribe");
+        }
+        return this.subscribeImpl(subject, queueName, handler);
+    }
 
     // Assumes the subj/queuename checks are done, does check for closed status
-    Dispatcher subscribeImpl(String subject, String queueName) {
+    Subscription subscribeImpl(String subject, String queueName, MessageHandler handler) {
         if (!this.running.get()) {
             throw new IllegalStateException("Dispatcher is closed");
         }
-        
+
         if (this.isDraining()) {
             throw new IllegalStateException("Dispatcher is draining");
         }
 
-        NatsSubscription sub = subscriptions.get(subject);
+        // If the handler is null, then we use the default handler, which will not allow
+        // duplicate subscriptions to exist.
+        if (handler == null) {
+            NatsSubscription sub = this.subscriptionsUsingDefaultHandler.get(subject);
 
-        if (sub == null) {
-            sub = connection.createSubscription(subject, queueName, this);
-            NatsSubscription actual = subscriptions.putIfAbsent(subject, sub);
-            if (actual != null) {
-                this.connection.unsubscribe(sub, -1); // Could happen on very bad timing
+            if (sub == null) {
+                sub = connection.createSubscription(subject, queueName, this);
+                NatsSubscription actual = this.subscriptionsUsingDefaultHandler.putIfAbsent(subject, sub);
+                if (actual != null) {
+                    this.connection.unsubscribe(sub, -1); // Could happen on very bad timing
+                }
             }
-        }
 
-        return this;
+            return sub;
+        } else {
+            NatsSubscription sub = connection.createSubscription(subject, queueName, this);
+            this.subscriptionsWithHandlers.put(sub.getSID(), sub);
+            this.subscriptionHandlers.put(sub.getSID(), handler);
+            return sub;
+        }
     }
 
     public Dispatcher unsubscribe(String subject) {
         return this.unsubscribe(subject, -1);
     }
 
+    public Dispatcher unsubscribe(Subscription subscription) {
+        return this.unsubscribe(subscription, -1);
+    }
+
     public Dispatcher unsubscribe(String subject, int after) {
         if (!this.running.get()) {
-            throw new IllegalStateException("Dispatcfher is closed");
+            throw new IllegalStateException("Dispatcher is closed");
         }
 
         if (isDraining()) { // No op while draining
@@ -207,8 +285,37 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
         if (subject == null || subject.length() == 0) {
             throw new IllegalArgumentException("Subject is required in unsubscribe");
         }
+
+        NatsSubscription sub = this.subscriptionsUsingDefaultHandler.get(subject);
+
+        if (sub != null) {
+            this.connection.unsubscribe(sub, after); // Connection will tell us when to remove from the map
+        }
+
+        return this;
+    }
+
+    public Dispatcher unsubscribe(Subscription subscription, int after) {
+        if (!this.running.get()) {
+            throw new IllegalStateException("Dispatcher is closed");
+        }
+
+        if (isDraining()) { // No op while draining
+            return this;
+        }
+
+        if (subscription.getDispatcher() != this) {
+            throw new IllegalStateException("Subscription is not managed by this Dispatcher");
+        }
+
+        // We can probably optimize this path by adding getSID() to the Subscription interface.
+        if (!(subscription instanceof NatsSubscription)) {
+            throw new IllegalArgumentException("This Subscription implementation is not known by Dispatcher");
+        }
         
-        NatsSubscription sub = subscriptions.get(subject);
+        NatsSubscription ns = ((NatsSubscription) subscription);
+        // Grab the NatsSubscription to verify we weren't given a different manager's subscription.
+        NatsSubscription sub = this.subscriptionsWithHandlers.get(ns.getSID());
 
         if (sub != null) {
             this.connection.unsubscribe(sub, after); // Connection will tell us when to remove from the map
@@ -218,7 +325,10 @@ class NatsDispatcher extends NatsConsumer implements Dispatcher, Runnable {
     }
 
     void sendUnsubForDrain() {
-        this.subscriptions.forEach((id, sub)->{
+        this.subscriptionsUsingDefaultHandler.forEach((id, sub)->{
+            this.connection.sendUnsub(sub, -1);
+        });
+        this.subscriptionsWithHandlers.forEach((sid, sub)->{
             this.connection.sendUnsub(sub, -1);
         });
     }