+++ src/test/java/io/nats/client/impl/DrainTests.java
@@ -90,7 +90,8 @@ public class DrainTests {
                 }
             });
             d.subscribe("draintest");
-            subCon.flush(Duration.ofSeconds(1)); // Get the sub to the server
+            d.subscribe("draintest", (msg) -> { count.incrementAndGet(); });
+            subCon.flush(Duration.ofSeconds(5)); // Get the sub to the server
 
             pubCon.publish("draintest", null);
             pubCon.publish("draintest", null);
@@ -99,10 +100,10 @@ public class DrainTests {
 
             // Drain will unsub the dispatcher, only messages that already arrived
             // are there
-            CompletableFuture<Boolean> tracker = d.drain(Duration.ofSeconds(4));
+            CompletableFuture<Boolean> tracker = d.drain(Duration.ofSeconds(8));
 
-            assertTrue(tracker.get(5, TimeUnit.SECONDS)); // wait for the drain to complete
-            assertEquals(count.get(), 2); // Should get both
+            assertTrue(tracker.get(10, TimeUnit.SECONDS)); // wait for the drain to complete
+            assertEquals(count.get(), 4); // Should get both, two times.
             assertFalse(d.isActive());
             assertEquals(((NatsConnection) subCon).getConsumerCount(), 0);
         }